+++ R/grd.R
@@ -8,14 +8,19 @@
 #' presented in \insertCite{ch19;textual}{superb}.
 #'
 #' @param RenameDV provide a name for the dependent variable (default DV) 
-#' @param SubjectsPerGroup indicates the number of simulated scores per group (default 100 in each group) 
-#' @param BSFactors a string indicating the between-subject factor(s) with, between parenthesis, the number of levels or the list of level names. Multiple factors are separated with a colon ":"
-#' @param WSFactors a string indicating the within-subject factor(s) in the same format as the between-subject factors
+#' @param SubjectsPerGroup indicates the number of simulated scores per 
+#'      group (default 100 in each group) 
+#' @param BSFactors a string indicating the between-subject factor(s) 
+#'      with, between parenthesis, the number of levels or the list of 
+#'      level names. Multiple factors are separated with a colon ":" or 
+#'      enumerated in a vector of strings.
+#' @param WSFactors a string indicating the within-subject factor(s) in 
+#'      the same format as the between-subject factors
 #' @param Effects a list detailing the effects to apply to the data
-#' @param Population a list providing the population characteristics (default is a normal distribution with a mean of 0 and standard deviation of 1)
-#' @param Contaminant a list providing the contaminant characteristics and the proportion of contaminant (default 0)
-#' @param Debug a boolean to produced detailed debugging information (default FALSE); useful for the maintainer only really
-#' @param Summary a boolean printing a summary of the design (default FALSE)
+#' @param Population a list providing the population characteristics 
+#'      (default is a normal distribution with a mean of 0 and standard deviation of 1)
+#' @param Contaminant a list providing the contaminant characteristics 
+#'      and the proportion of contaminant (default 0)
 #'
 #'
 #' @return a data.frame with the simulated scores.
@@ -93,9 +98,7 @@ GRD <- function(
     WSFactors = "",
     Effects = list(),
     Population  = list(mean = 0, stddev = 1, rho = 0, scores="rnorm(1, mean = GM, sd = STDDEV)"),
-    Contaminant = list(mean = 0, stddev = 1, rho = 0, scores="rnorm(1, mean = CGM, sd = CSTDDEV)", proportion = 0),
-    Debug = FALSE, 
-    Summary = FALSE
+    Contaminant = list(mean = 0, stddev = 1, rho = 0, scores="rnorm(1, mean = CGM, sd = CSTDDEV)", proportion = 0)
 ){
     #####################################################
     # STEP 0: Load required library
@@ -115,7 +118,7 @@ GRD <- function(
         BSList <- grdUnpacker(BSFactors, ':', 'BS')
     }
     ngroups <- prod(unlist(lapply(BSList,length)))
-    runDebug(Debug, "processing BSFactors", 
+    runDebug("GRD:1.1", "GRD:1.1: Processing BSFactors", 
         c("BSList","ngroups"),list(BSList,ngroups))
 
     # determining repeated measures
@@ -126,7 +129,7 @@ GRD <- function(
         WSList <- unpacked
     }
     nreplic <- prod(unlist(lapply(WSList,length)))
-    runDebug(Debug, "processing WSFactors", 
+    runDebug("GRD:1.2", "GRD:1.2: Processing WSFactors", 
         c("WSList","nreplic"),list(WSList,nreplic))
 
     # validating groups sizes
@@ -144,11 +147,11 @@ GRD <- function(
         sum(SubjectsPerGroup) 
     }
     rows = subj * nreplic
-    runDebug(Debug, "processing matrix size", 
+    runDebug("GRD:1.3", "GRD:1.3: Processing matrix size", 
         c("BSnames","WSnames","facnames","cols","rows","subj"   ),
         list(BSnames,WSnames,facnames,cols,rows,subj))
 
-    if (Summary) {
+    if ('summary' %in% getOption("superb.debug") ) {
         grdShowDesign(BSList, WSList, ngroups, nreplic, subj, SubjectsPerGroup)
     }
 
@@ -184,6 +187,10 @@ GRD <- function(
             data[,1+length(BSList)+i] = unlist(lapply(design[,i],rep,times=subj))
         }
     }
+
+    runDebug("GRD:2", "GRD:2: Processing base matrix", 
+        c("dataMatrix"),
+        list(data))
 
   
     #####################################################
@@ -244,6 +251,19 @@ GRD <- function(
             CDIST   <- "mvrnorm(1, mu = CGM, Sigma = CSIGMA)" }
     }
 
+    runDebug("GRD:3", "GRD:3: Processing population", 
+        c("GM","STDDEV","RHO","DIST","CGM","CSTDDEV","CRHO","CDIST","PROP"),
+        list(GM,STDDEV,RHO,DIST,CGM,CSTDDEV,CRHO,CDIST,PROP))
+    if ((RHO != 0)||(CRHO != 0)) {
+        runDebug("GRD.rho", "GRD.rho: Processing multivariate population", 
+            c("SIGMA","CSIGMA"),
+            list(SIGMA,CSIGMA))
+    }
+
+
+    #####################################################
+    # STEP 4: Getting effect definitions
+    #####################################################
 
     # generating a function F0 (population) and C0 (contaminant) from this
     neffects=0
@@ -253,30 +273,20 @@ GRD <- function(
     fctstrb <- paste("C0 <- function(", allfactors, ") { ", CDIST, " }", sep = "" )
     eval(parse(text=fctstrb))
 
-    runDebug(Debug, "processing population", 
-        c("GM","STDDEV","RHO","DIST","fctstra","CGM","CSTDDEV","CRHO","CDIST","PROP","fctstrb"),
-        list(GM,STDDEV,RHO,DIST,fctstra,CGM,CSTDDEV,CRHO,CDIST,PROP,fctstrb))
-    if ((RHO != 0)||(CRHO != 0)) {
-        runDebug(Debug, "processing multivariate population", 
-            c("SIGMA","CSIGMA"),
-            list(SIGMA,CSIGMA))
-    }
-
-
-    #####################################################
-    # STEP 4: Getting effect definitions
-    #####################################################
-    # make functions that mapped to the data returns the effect  
+    # make functions that when mapped to the data returns the effect  
     if (length(Effects) > 0 ) { 
         for (i in 1:length(Effects)) {
             neffects = neffects + 1
             # make effect and run it.
             fctstr = grdMakeEffect(neffects, names(Effects)[i], Effects[[i]], 
-                c(BSList, WSList), data, allfactors, Debug
+                c(BSList, WSList), data, allfactors
             )
             eval(parse(text=fctstr))
         }
     }
+    runDebug("GRD:4", "GRD:4: Processing the effects", 
+        c("fctstrF","fctstrC"),
+        list(fctstra,fctstrb))
   
   
     ############################################
@@ -284,10 +294,10 @@ GRD <- function(
     ############################################
     # runs F0 effect
     if (multivariate) {
-        temp1    <- function(args) { do.call(what="F0",args=as.list(args),quote=FALSE)}
+        temp1    <- function(args) { do.call(what="F0", args=as.list(args), quote=FALSE)}
         deltaDV  <- as.vector( t(apply(data[1:subj,], 1, temp1)) ) 
     } else {
-        temp1    <- function(args) { do.call(what="F0",args=as.list(args),quote=FALSE)}
+        temp1    <- function(args) { do.call(what="F0", args=as.list(args), quote=FALSE)}
         deltaDV  <- apply(data, 1, temp1)    
     }
     data$DV <- data$DV + deltaDV
@@ -340,7 +350,6 @@ GRD <- function(
 
 
 
-
 ######################################################################################
 #' @title effect description
 #'
@@ -367,9 +376,8 @@ GRD <- function(
 #' @export Rexpression
 #'
 
-
 ##################################################################   
-# Aesthetic private functions
+# Aesthetic functions
 ##################################################################   
 
 # These provide nicer inputs to grd
@@ -383,7 +391,7 @@ Rexpression <- function(s)   { c("-96", paste(c(s),collapse="") ) }
 # Subsidiary private functions: grdMakeEffect; grdL2W; grdDepacker; grdUnpacker;
 ##################################################################   
 
-grdMakeEffect <- function (fnumber, name, details, WBSList, data, allfactors, Debug) {
+grdMakeEffect <- function (fnumber, name, details, WBSList, data, allfactors ) {
 
     if (details[1] == "-96") {
         # this is an R expression
@@ -421,7 +429,7 @@ grdMakeEffect <- function (fnumber, name, details, WBSList, data, allfactors, De
                  "switch( ",swit,subs,") }", sep = "" )
     }
     
-    runDebug(Debug, paste("Effect functions",fnumber), 
+    runDebug("GRD.Effect", paste("GRD.Effect: Processing effect declaration ", fnumber), 
         c("eff","pfaclevel","fctstr"),
         list(eff,pfaclevel,fctstr))
 
@@ -431,7 +439,7 @@ grdMakeEffect <- function (fnumber, name, details, WBSList, data, allfactors, De
 
 grdL2W <- function (data, within, dv = "DV") {
     # the function grdL2W taken from Navarro, "lsr" library longToWide.
-    # The original function is a bit different so I made a few minor changes
+    # The original function is a bit different so I made minor changes.
     idvar <- setdiff(names(data), c(within, dv))
     if (length(within) > 1) {
         collapsed.treatments <- apply(as.matrix(data[, within]), 
@@ -498,8 +506,12 @@ grdUnpacker = function(string, sep_across, defaultfactorname) {
             # Identify levels
             lvls <- gsub(".*\\((.*)\\).*", "\\1", varnlvls)
             lvls <- unlist(strsplit(lvls,","))
-              
-            #If it is a single number, assume it is the number of levels, not its name
+                
+            # If the two levels are numeric, e.g., (1,4) convert to numeric
+            if ((length(lvls) == 2) & all(suppressWarnings(!is.na(as.numeric(lvls)))))
+                lvls <- as.numeric(lvls[1]):as.numeric(lvls[2])
+
+            # If it is a single number, assume it is the number of levels, not the level name
             if (length(lvls) == 1 && suppressWarnings(!is.na(as.numeric(lvls))))  {
                 lvls <- 1:lvls
             }