+++ tests/testthat/test_superbPlot.R
@@ -50,26 +50,33 @@ test_that("PRELIMINARY TESTS (4/4)", {
 #########################################
 
 test_that("test 1a: 3 groupes inpépendants", { 
+    options(superb.debug = c('none'))
     dta1a <- GRD( BSFactors = "Group(3)", Population = list( mean=10, stddev = 5) )
     # write.table(dta1a, file = "test1a.dat", sep = "\t", col.names = FALSE)
     p <- superbPlot(dta1a, BSFactor = "Group", variables = "DV",
       statistic = "mean", errorbar = "SE", plotStyle="line")
 
     expect_equal( "ggplot" %in% class(p), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
 test_that("test 1b: factorielle à grps indépendants; 3 x 2", {
+    options(superb.debug = c('none'))
     dta1b <- GRD( BSFactors = "Group(3): Sex(2)", Population = list( mean=10, stddev = 5))
     # write.table(dta1b, file = "test1b.dat", sep = "\t", col.names = FALSE)
     p <- superbPlot(dta1b, BSFactor = c("Group","Sex"), variables = "DV",
       statistic = "mean", errorbar = "SE" )
 
     expect_equal( "ggplot" %in% class(p), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
 test_that("test 2a: 1 facteur à 3 mesures répétées; (3)", {
+    options(superb.debug = c('design','warnings'))
     dta2a <- GRD( WSFactors = "Moment(3)", SubjectsPerGroup = 5, Population = list( mean=10, stddev = 5))
     # write.table(dta2a, file = "test2a.dat", sep = "\t", col.names = FALSE)
     expect_warning( p <- superbPlot(dta2a, WSFactor = "moment(3)", 
@@ -79,6 +86,8 @@ test_that("test 2a: 1 facteur à 3 mesures répétées; (3)", {
     ))
 
     expect_equal( "ggplot" %in% class(p), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -100,6 +109,7 @@ test_that("test 2b: 2 facteurs à mesures répétées; (3 x 2)", {
 
 
 test_that("test 3: schème mixte; 3 x (3)", {
+    options(superb.debug = c('design','warnings'))
     dta3 <- GRD( BSFactors = "Group(3)", WSFactors = "Moment(3)", 
       SubjectsPerGroup = 5, Population = list( mean=10, stddev = 5),
       Effects = list("Moment" = slope(5))
@@ -112,10 +122,13 @@ test_that("test 3: schème mixte; 3 x (3)", {
     ))
 
     expect_equal( "ggplot" %in% class(p), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
 test_that("test 4a: schème à trois facteurs, 2 étant between  3 x 3 x (3)", {
+    options(superb.debug = c('design','warnings'))
     dta4a <- GRD( BSFactors = "Group(3) : Dose(3)", WSFactors = "Moment(3)", 
       SubjectsPerGroup = 4, Population = list( mean=10, stddev = 5),
       Effects = list("Moment" = slope(5))
@@ -128,14 +141,16 @@ test_that("test 4a: schème à trois facteurs, 2 étant between  3 x 3 x (3)", {
       factorOrder = c("Dose","Group","Moment"), showPlot = T)
 
     expect_equal( "ggplot" %in% class(p), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
     
   
 test_that("test 5a: schème à quatre facteurs; 5 x 4 (3 x 2)", {
     options(superb.debug = "none") # to supress design confirmation; unneeded in tests
     dta5a <- GRD( BSFactors = "Group(5) : Dose(4)", WSFactors = "Moment(3):Hand(2)", 
-      Summary=FALSE, Population = list( mean=10, stddev = 5, rho = .90),
-      Effects = list("Moment" = slope(5), "Hand" = slope(10)) )
+        Population = list( mean=10, stddev = 5, rho = .90),
+        Effects = list("Moment" = slope(5), "Hand" = slope(10)) )
     # write.table(dta5a, file = "test5a.dat", sep = "\t", col.names = FALSE)
     p <- superbPlot(dta5a, plotStyle="line",
         WSFactor = c("Moment(3)","Hand(2)"), 
@@ -155,12 +170,15 @@ test_that("test 5a: schème à quatre facteurs; 5 x 4 (3 x 2)", {
 #########################################
 
 test_that("test 6: Some data", {
+    options(superb.debug = c('none'))
     dta6 <- GRD( WSFactors = "Moment(3):Hand(2)",
         Effects = list("Moment" = slope(5), "Hand" = slope(3)),
         SubjectsPerGroup = 6,
         Population = list (mean = 20, stddev = 5, rho = 0.8) )
 
     expect_output( str(dta6), "data.frame")
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -184,7 +202,8 @@ test_that("test 6a: factorOrder", {
 
     expect_equal( "ggplot" %in% class(p1), TRUE)
     expect_equal( "ggplot" %in% class(p2), TRUE)
-    options(superb.debug = c("design","warnings") ) # restores to default
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -216,7 +235,8 @@ test_that("test 6b: adjustments CA vs CM vs LM", {
     expect_equal( "ggplot" %in% class(p1), TRUE)
     expect_equal( "ggplot" %in% class(p2), TRUE)
     expect_equal( "ggplot" %in% class(p3), TRUE)
-    options(superb.debug = c("design","warnings") ) # restores to default
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -247,7 +267,8 @@ test_that("test 6c: statistics of central tendency mean, median and gmean", {
     expect_equal( "ggplot" %in% class(p1), TRUE)
     expect_equal( "ggplot" %in% class(p2), TRUE)
     expect_equal( "ggplot" %in% class(p3), TRUE)
-    options(superb.debug = c("design","warnings") ) # restores to default
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -272,7 +293,8 @@ test_that("test 6d: statistics of dispersion sd and MAD", {
 
     expect_equal( "ggplot" %in% class(p1), TRUE)
     expect_equal( "ggplot" %in% class(p2), TRUE)
-    options(superb.debug = c("design","warnings") ) # restores to default
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -325,6 +347,7 @@ test_that("test 6g: adding ggplot options to the error bars, to the points (bis)
 #########################################
 
 test_that("Explorations for ICC", {
+    options(superb.debug = c('design','warnings'))
     library(gridExtra)
     dta99 <- GRD( WSFactors = "Moment(3)", SubjectsPerGroup = 15, 
         Population = list( mean=20, stddev = 5),
@@ -353,6 +376,8 @@ test_that("Explorations for ICC", {
 
     expect_equal( "ggplot" %in% class(noncluster), TRUE)
     expect_equal( "ggplot" %in% class(yescluster), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -362,6 +387,7 @@ test_that("Explorations for ICC", {
 ########################################################
 
 test_that("Verifying CA and popSize ", {
+    options(superb.debug = c('design','warnings'))
     library(gridExtra)
     dta3 <- GRD( BSFactors = "Group(2)", WSFactors = "Moment(3)", 
       SubjectsPerGroup = 5, Population = list (mean = 20, stddev = 5, rho = 0.8),
@@ -382,6 +408,8 @@ test_that("Verifying CA and popSize ", {
 
     expect_equal( "ggplot" %in% class(p1), TRUE)
     expect_equal( "ggplot" %in% class(p2), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -390,6 +418,7 @@ test_that("Verifying CA and popSize ", {
 #########################################
 
 test_that("Testing pre and post processing", {
+    options(superb.debug = c('design','warnings'))
     library(ggplot2)
     library(gridExtra)
     dta9 <- GRD( WSFactors = "Moment(3)", SubjectsPerGroup = 5, 
@@ -451,6 +480,8 @@ test_that("Testing pre and post processing", {
     expect_equal( "ggplot" %in% class(altlm), TRUE)
     expect_equal( "ggplot" %in% class(truecmvslm), TRUE)
     expect_equal( "ggplot" %in% class(altcmvslm), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
 
 
@@ -459,6 +490,7 @@ test_that("Testing pre and post processing", {
 #########################################
 
 test_that("Many tests with TMB1964r", {
+    options(superb.debug = c('none'))
     library(ggplot2)
     options(superb.debug = "none")
     mee = TMB1964r[TMB1964r$Language == "English"|TMB1964r$Language == "French",]
@@ -520,5 +552,88 @@ test_that("Many tests with TMB1964r", {
     expect_equal( "ggplot" %in% class(plt4), TRUE)
     expect_equal( "ggplot" %in% class(plt5), TRUE)
 
-
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
 })
+
+
+test_that("Heterogeneous variances", {
+    options(superb.debug = c('none'))
+    dta <- GRD(
+        BSFactors = "dif(3) : grp (2)", 
+        WSFactors="day(1,2)",
+        SubjectsPerGroup = 3,
+        Population=list(
+            mean = 100, 
+            scores = "rnorm(1, mean = GM, sd = 100 * (grp-1) +0.1)"
+        )
+    )
+    options(superb.debug = c('design','warnings'))
+    expect_warning( 
+        superbPlot(dta,
+            BSFactor = c("dif","grp"),
+            WSFactor = "day(2)",
+            variables = c("DV.1","DV.2"),
+            adjustment = list( purpose = "difference") )
+    )
+
+    plt <- superbPlot(dta,
+        BSFactor = c("dif","grp"),
+        WSFactor = "day(2)",
+        variables = c("DV.1","DV.2"),
+        adjustment = list( purpose = "tryon") )
+
+    expect_equal( "ggplot" %in% class(plt), TRUE )
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
+})
+
+
+test_that("Tryon vs. difference", {
+    options(superb.debug = c('design','warnings'))
+    #lets have data with massively heterogeneous variances
+    dta <-GRD( BSFactors="grp(3)",
+        RenameDV = "score",
+        Population=list(
+            mean = 100, 
+            scores = "rnorm(1, mean = GM, sd = 10 * grp)"
+                        ),
+        SubjectsPerGroup = 25,
+        Effects = list("grp" = slope(15) ) 
+    )
+    # do the plot twice with transparant background
+    expect_warning(plt1 <- superbPlot(dta, 
+        BSFactor = "grp", plotStyle="line",
+        variables = "score",  
+        errorbarParams = list(color="blue",position = position_nudge(-0.1) ),
+        adjustments = list( purpose = "difference") ) + 
+     labs(title="(blue) Difference-adjusted 95% confidence intervals\n(red) Tryon 95% confidence intervals") +
+     coord_cartesian( ylim = c(65,135) ) + 
+     theme(panel.background = element_rect(fill = "transparent"),
+           plot.background = element_rect(fill = "transparent", color = "white")) 
+    )
+    plt2 <- superbPlot(dta, 
+        BSFactor = c("grp"), plotStyle="line",
+        variables = "score",  
+        errorbarParams = list(color="red",position = position_nudge(+0.1) ),
+        adjustments = list( purpose = "tryon") ) +
+     labs(title="(blue) Difference-adjusted 95% confidence intervals\n(red) Tryon 95% confidence intervals") +
+     coord_cartesian( ylim = c(65,135) ) + 
+     theme(panel.background = element_rect(fill = "transparent"),
+           plot.background = element_rect(fill = "transparent", color = "white"))
+    # transform the plots in manipulable objets
+    plt1g <- ggplotGrob(plt1)
+    plt2g <- ggplotGrob(plt2)
+    # put the two grob onto an empty ggplot (as the positions are the same, they will be overlayed)
+    plt <- ggplot() + 
+        annotation_custom(grob=plt1g) + 
+        annotation_custom(grob=plt2g)
+    expect_equal( "ggplot" %in% class(plt), TRUE)
+    # restores default information
+    options(superb.debug = c('design','warnings','summary'))
+})
+
+
+
+
+